# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prime Wave API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /semesters/{id}:
    get:
      operationId: getSemester
      tags: [semesters]
      summary: Get semester with modules
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Semester
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Semester"
  /semesters/{id}/modules:
    get:
      operationId: listModules
      tags: [semesters]
      summary: List modules for a semester
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
  /students/{studentId}/semesters/{semesterId}/progress:
    get:
      operationId: getSemesterProgress
      tags: [semesters]
      summary: Get student progress in a semester
      parameters:
        - in: path
          name: studentId
          required: true
          schema: { type: integer }
        - in: path
          name: semesterId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemesterProgress"

  # ─── TOPICS ─────────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: studentId
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ─── CERTIFICATES ───────────────────────────────────────────────────────────
...
  /portfolio/{studentId}:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get student portfolio
      parameters:
        - in: path
          name: studentId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"

  # ─── LEADERBOARD ────────────────────────────────────────────────────────────
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job postings
      parameters:
        - in: query
          name: companyId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/candidates:
    get:
      operationId: getJobCandidates
      tags: [jobs]
      summary: Get AI-matched candidates for a job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Matched candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateMatch"

  # ─── COMPANIES ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Semester:
      type: object
      required: [id, number, title, description, creditsRequired, status]
      properties:
        id:              { type: integer }
        number:          { type: integer }
        title:           { type: string }
        description:     { type: string }
        creditsRequired: { type: integer }
        status:          { type: string }
        modules:
          type: array
          items:
            $ref: "#/components/schemas/Module"
    Module:
      type: object
      required: [id, semesterId, title, description, order, status]
      properties:
        id:          { type: integer }
        semesterId:  { type: integer }
        title:       { type: string }
        description: { type: string }
        order:       { type: integer }
        status:      { type: string }
        topicsCount: { type: integer }
        score:       { type: ["number", "null"] }
    SemesterProgress:
      type: object
      required: [semesterId, studentId, completedModules, totalModules, completedTopics, totalTopics, overallScore, status]
      properties:
        semesterId:        { type: integer }
        studentId:         { type: integer }
        completedModules:  { type: integer }
        totalModules:      { type: integer }
        completedTopics:   { type: integer }
        totalTopics:       { type: integer }
        overallScore:      { type: number }
        status:            { type: string }
        creditsEarned:     { type: ["integer", "null"] }

    # ─── TOPIC ────────────────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, studentId, title, type, status, createdAt]
      properties:
        id:          { type: integer }
        studentId:   { type: integer }
        moduleId:    { type: ["integer", "null"] }
        semesterId:  { type: ["integer", "null"] }
        title:       { type: string }
        description: { type: ["string", "null"] }
        type:        { type: string }
        status:      { type: string }
        githubUrl:   { type: ["string", "null"] }
        score:       { type: ["number", "null"] }
        techStack:
          type: array
          items: { type: string }
        createdAt:   { type: string }
        submittedAt: { type: ["string", "null"] }
    ProjectInput:
      type: object
      required: [studentId, title, type]
      properties:
        studentId:   { type: integer }
        moduleId:    { type: integer }
        semesterId:  { type: integer }
        title:       { type: string }
        description: { type: string }
        type:        { type: string }
        techStack:
          type: array
          items: { type: string }
    ProjectUpdate:
      type: object
      properties:
        title:       { type: string }
        description: { type: string }
        githubUrl:   { type: string }
        status:      { type: string }
        techStack:
          type: array
          items: { type: string }

    # ─── CERTIFICATE ──────────────────────────────────────────────────────────
    Certificate:
      type: object
      required: [id, studentId, type, title, issuedAt, creditsAwarded, verificationCode]
      properties:
        id:               { type: integer }
        studentId:        { type: integer }
        type:             { type: string }
        title:            { type: string }
        semesterNumber:   { type: ["integer", "null"] }
        issuedAt:         { type: string }
        creditsAwarded:   { type: integer }
        score:            { type: ["number", "null"] }
        verificationCode: { type: string }
        qrCodeUrl:        { type: ["string", "null"] }

    # ─── PORTFOLIO ────────────────────────────────────────────────────────────
    Portfolio:
      type: object
      required: [studentId, ppsScore, creditsEarned, industryReadiness, semestersCompleted]
      properties:
        studentId:           { type: integer }
        studentName:         { type: string }
        university:          { type: string }
        ppsScore:            { type: number }
        creditsEarned:       { type: integer }
        industryReadiness:   { type: number }
        semestersCompleted:  { type: integer }
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        skills:
          type: array
          items: { type: string }
        githubUrl:           { type: ["string", "null"] }
        linkedinUrl:         { type: ["string", "null"] }

    # ─── LEADERBOARD ─────────────────────────────────────────────────────────
...
    Job:
      type: object
      required: [id, companyId, companyName, title, description, status, createdAt]
      properties:
        id:           { type: integer }
        companyId:    { type: integer }
        companyName:  { type: string }
        title:        { type: string }
        description:  { type: string }
        requiredSkills:
          type: array
          items: { type: string }
        minPpsScore:  { type: ["number", "null"] }
        minCredits:   { type: ["integer", "null"] }
        location:     { type: ["string", "null"] }
        type:         { type: string }
        status:       { type: string }
        applicantsCount: { type: integer }
        createdAt:    { type: string }
    JobInput:
      type: object
      required: [companyId, title, description]
      properties:
        companyId:    { type: integer }
        title:        { type: string }
        description:  { type: string }
        requiredSkills:
          type: array
          items: { type: string }
        minPpsScore:  { type: number }
        minCredits:   { type: integer }
        location:     { type: string }
        type:         { type: string }
    CandidateMatch:
      type: object
      required: [studentId, studentName, matchScore, ppsScore, creditsEarned, semesterNumber]
      properties:
        studentId:        { type: integer }
        studentName:      { type: string }
        avatar:           { type: ["string", "null"] }
        university:       { type: string }
        matchScore:       { type: number }
        ppsScore:         { type: number }
        creditsEarned:    { type: integer }
        semesterNumber:   { type: integer }
        industryReadiness: { type: number }
        skills:
          type: array
          items: { type: string }
        githubUrl:        { type: ["string", "null"] }
        portfolioUrl:     { type: ["string", "null"] }
        skillsMatchPct:   { type: number }
...