# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Prime Wave API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Module {
  id: number;
  semesterId: number;
  title: string;
  description: string;
  order: number;
  status: string;
  topicsCount?: number;
  /** @nullable */
  score?: number | null;
}